% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissertation-package.R
\docType{package}
\name{irtc-package}
\alias{irtc}
\alias{irtc-package}
\title{irtc}
\description{
R package supporting a dissertation provisioned via dissertation.ai.
Currently contains scaffolding only. Add your methodological R functions
to the \verb{R/} directory and roxygen-document them; tests go in
\verb{tests/testthat/}.
}
\details{
The Quarto book at \verb{ui/www/} is the thesis source. The
\code{build-thesis.yml} GitHub Action renders it to PDF on every push.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dbetebenner/irtc}
  \item Report bugs at \url{https://github.com/dbetebenner/irtc/issues}
}

}
\author{
\strong{Maintainer}: Damian Betebenner \email{dbetebenner@gmail.com}

Authors:
\itemize{
  \item Damian Betebenner \email{dbetebenner@gmail.com}
}

}
\keyword{internal}
