% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissertation-package.R
\name{package_info}
\alias{package_info}
\title{Package metadata (stub function)}
\usage{
package_info()
}
\value{
A list with package name + version.
}
\description{
Replaced at provision time + extended by \code{dataimago::ai()} based on your
\code{dataimago-spec.yaml}.
}
\examples{
package_info()
}
