% Copyright (C) 2026 WEIAN DATA TECH (Beijing) Co., Ltd.
% SPDX-License-Identifier: GPL-2.0-or-later
\name{IRTC-package}
\alias{IRTC-package}
\alias{IRTC}
\docType{package}
\title{
Marginal Maximum Likelihood Estimation for Item Response Models
}
\description{
A self-contained marginal maximum likelihood estimation package for
unidimensional and multidimensional item response models. It supports
Rasch / 1PL, partial credit, rating scale, 2PL and generalized partial
credit models, latent regression, multiple groups, case weights, and a
bounded-memory streaming engine for large simple-structure models.
}
\details{
The main function is \code{\link{irtc.mml}}. It returns an object of class
\code{irtc} with \code{summary}, \code{print}, \code{logLik} and
\code{anova} methods.
}
\author{
Kunxiang Ma and WEIAN DATA TECH (Beijing) Co., Ltd.
}
\seealso{
\code{\link{irtc.mml}}
}
\keyword{package}
