% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-result.R
\name{model_result_to_json}
\alias{model_result_to_json}
\title{Serialize a model result to JSON}
\usage{
model_result_to_json(result)
}
\arguments{
\item{result}{An \code{irtc_model_result}.}
}
\value{
A JSON string.
}
\description{
Serialize a model result to JSON
}
\seealso{
Other model results:
\code{\link[=model_result_from_json]{model_result_from_json()}},
\code{\link[=new_model_result]{new_model_result()}}
}
\concept{model results}
